Packet sniffing added and new Affixes structure.
Packets now block out the name from a PlayerInfoPacket and give the
DisplayName as the MC name (so tab menu works correctly). Clear name
means name is entirely set with team prefix/suffix

Affixes show the way I want to rewrite this piece of code as it was
originally a huge mess.